working on new specs
