+++ src/test/scala/lilo/LiloSourceSpec.scala
@@ -1,0 +1,10 @@
+package lilo
+
+import org.specs2.mutable.Specification
+import org.specs2.mock.Mockito
+
+class LiloSourceSpec extends Specification with Mockito {
+  
+  
+
+}
